+++ CHANGELOG.md
@@ -2,22 +2,45 @@
 
 ## Unreleased
 - Allow definition of a default queue.
-- Change command run_djangohuey to djangohuey.
 - Allow definition of queues by django app.
 - Allow definition of specific queues file path.
 
 ## 1.0.0 - 
-**Note** This release contains breaking changes, see them below with the migration instructions.
+**Note:** This release contains breaking changes, see them below with the migration instructions.
 
 
 ### Changed
-- HUEYS django setting renamed to DJANGO_HUEY
+- HUEYS django setting renamed to DJANGO_HUEY.
+- Change command run_djangohuey to djangohuey.
 
 
-### Breaaking Changes
-- HUEYS django setting renamed to DJANGO_HUEY
-**Migration example**
+### Breaking Changes
+#### *HUEYS django setting renamed to DJANGO_HUEY.*
 
+##### Migration steps: 
+In *django settings* change:
+```python
+HUEYS = {...} #Queues definition
+```
+
+to the following:
+```python
+DJANGO_HUEY = {...} #Queues definition
+```
+
+#### *Change command run_djangohuey to djangohuey.*
+
+##### Migration steps: 
+Before this release you run the queue consumer with the following command:
+```bash
+python manage.py run_djangohuey --queue first
+```
+
+Now you run:
+```bash
+python manage.py djangohuey --queue first
+```
+---
 
 ## 0.2.0 - 2021-04-18